Added a lot of things in this commit
Created class Centroid to model the output of kmeans.
Changed value of Parameters.PIVOTS_FILE to the one we use for storing
centroids.
Added some comments here and there in the code.
Added the code to create posting lists, this means we create a list of
centroids and for each image we compute the distance from each keypoint
of the image to the centroid.