Emergency commit
We solved OutOfMemory error and changed the way we get features from
each image.
We added the class for reading the JSON and added a method in Output for
creating the web page to show to the user.
We changed some parameter in order to have consistency (from String to
File)
Added field of the Json to retrieve